etl/requirements.txt
requests>=2.32
urllib3>=2.2
python-dotenv>=1.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dateutil>=2.9
psycopg[binary]>=3.2
